+++ packages/labs/src/index.ts
@@ -11,5 +11,4 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-export * from './callbacks/callbacks.js';
 export * from './controllers/updating-controller.js';