+++ packages/lit-html/src/test/directives/live_test.ts
@@ -15,7 +15,6 @@
 import {live} from '../../directives/live.js';
 import {html, noChange, nothing, render} from '../../lit-html.js';
 import {assert} from '@esm-bundle/chai';
-import '../polyfills.js';
 
 class LiveTester extends HTMLElement {
   _x?: string;