+++ rollup-common.js
@@ -54,13 +54,16 @@ const reservedProperties = [
   // can share this key with dev lit-element platform-support which
   // imports it.
   '_handlesPrepareStyles',
-  // Note, reserved so that labs in dev mode can see these properties on
-  // updating-element in prod mode. Labs imports via 'updating-element' and
-  // therefore gets prod mode by default.
-  '_willUpdate',
-  '_didUpdate',
-  '_connectedCallback',
-  '_disconnectedCallback',
+  // TODO(kschaaf) TBD: lit-ssr required "private" fields (can be in
+  // crossPackagePropertyMangles once lit-ssr uses the rollup config)
+  // lit-html: AttributePart (used by render-lit-html)
+  '_commitValue',
+  // lit-html: Directive (used by render-lit-html)
+  '_resolve',
+  // updating-element: UpdatingElement (used by lit-element-renderer)
+  '_attributeToProperty',
+  // hydrate-support: LitElement (added by hydrate-support)
+  '_needsHydration',
 ];
 
 // Any private properties which we share between different _packages_ are
@@ -79,13 +82,10 @@ const crossPackagePropertyMangles = {
   _getTemplate: 'F',
   // lit-html: TemplateInstance
   _template: 'H',
-  // updating-element: UpdatingElement
-  _willUpdate: 'T',
-  _didUpdate: 'S',
-  _connectedCallback: 'U',
-  _disconnectedCallback: 'V',
   // lit-element: LitElement
   _renderOptions: 'W',
+  // lit-element: LitElement (used by hydrate-support)
+  _renderImpl: 'M',
 };
 
 const generateTerserOptions = (nameCache = null) => ({